added %DOU

+++ ft_printf/print_funs/print_f.c
@@ -46,13 +46,16 @@ char		*btoa(long double f, int precision, int hash)
 	char *temp;
 
 	ret = ft_strnew(3222);
+//	printf("(minus_zero = %Lf)\n", minus_zero);
+//	printf("(f = %Lf)\n", f);
 	if (f < 0 || f == minus_zero)
 	{
+//		printf("Kaka\n");
 		ret[0] = '-';
 		f = -f;
 	}
 
-	hard = (unsigned long long)(float)f;
+	hard = (unsigned long long)f;
 	temp = ft_itoa_ulong(hard, 10);
 	ret = ft_strcat(ret, temp);
 	free(temp);
@@ -61,8 +64,15 @@ char		*btoa(long double f, int precision, int hash)
 	precision = (precision == 0) ? 6 : precision;
 	while (precision-- > 0)
 	{
-		f = (f - hard) * 10;
+//		printf("(precision = %d)\n", precision);
+//		printf("(f1 = %Lf)\n", f);
+//		printf("(hard1 = %llu)\n", hard);
+		f = (f - hard) * 10.0;
+//		printf("(f2 = %Lf)\n", f);
+
+
 		hard = (unsigned long long)(float)f;
+//		printf("(hard1 = %llu)\n", hard);
 		if ((f - hard) * power_ten(precision + 1) > power_ten(precision + 1) - 5)
 		{
 			hard = (hard + 1) * power_ten(precision);
@@ -71,6 +81,7 @@ char		*btoa(long double f, int precision, int hash)
 		temp = ft_itoa_ulong(hard, 10);
 		ret = ft_strcat(ret, temp);
 		free(temp);
+//		printf("________________\n");
 	}
 	return (ret);
 }
